rename sep241util
