+++ sep241peakcalling.py
@@ -10,10 +10,10 @@ from tqdm.auto import tqdm
 from threadpoolctl import threadpool_limits
 import multiprocessing
 
-from dcbackend import logger, setup_logging
-from dcbackend import read_job_data, read_results, write_bed, detect_cores
-from dcbackend import check_length_distribution_flip
-from dcbackend import MissingData
+from sep241util import logger, setup_logging
+from sep241util import read_job_data, read_results, write_bed, detect_cores
+from sep241util import check_length_distribution_flip
+from sep241util import MissingData
 
 
 def parse_args():
