+++ sep241util.py
